User API

Index

MuseInference.MuseResult

MuseInference.TuringMuseProblem

MuseInference.muse

Contents

MuseInference.muse — Function

muse(prob::AbstractMuseProblem, θ₀; kwargs...)
muse!(result::MuseResult, prob::AbstractMuseProblem, [θ₀=nothing]; kwargs...)

Run the MUSE estimate. The muse! form resumes an existing result. If the muse form is used instead, θ₀ must give a starting guess for $\theta$.

See MuseResult for description of return value.

Optional keyword arguments:

rng — Random number generator to use. Taken from result.rng or Random.default_rng() if not passed.

z₀ — Starting guess for the latent space MAP.

maxsteps = 50 — Maximum number of iterations.

θ_rtol = 1e-2 — Error tolerance on $\theta$ relative to its standard deviation.

∇z_logLike_atol = 1e-2 — Absolute tolerance on the $z$-gradient at the MAP solution.

nsims = 100 — Number of simulations.

α = 0.7 — Step size for root-finder.

progress = false — Show progress bar.

pmap — Parallel map function.

regularize = identity — Apply some regularization after each step.

H⁻¹_like = nothing — Initial guess for the inverse Jacobian of $s^{\rm MUSE}(\theta)$

H⁻¹_update — How to update H⁻¹_like. Should be :sims, :broyden, or :diagonal_broyden.

broyden_memory = Inf — How many past steps to keep for Broyden updates.

checkpoint_filename = nothing — Save result to a file after each iteration.

get_covariance = false — Also call get_H and get_J to get the full covariance.

MuseInference.MuseResult — Type

Stores the result of a MUSE run. Can be constructed by-hand as MuseResult() and passed to any of the inplace muse!, get_J!, or get_H!.

Fields:

θ — The estimate of the $\theta$ parameters.

Σ, Σ⁻¹ — The approximate covariance of $\theta$ and its inverse.

H, J — The $H$ and $J$ matrices which form the covariance (see Millea & Seljak, 2021)

gs — The MAP gradient sims used to compute J.

Hs — The jacobian sims used to compute H.

dist — A Normal or MvNormal built from θ and Σ, for convenience.

history — Internal diagnostic info from the run.

rng — RNG used to generate sims for this run (so the same sims can be reused if resuming later).

time — Total Millisecond wall-time spent computing the result.

MuseInference.TuringMuseProblem — Type

TuringMuseProblem(model; params = (:θ,), autodiff = Turing.ADBACKEND)

Wrap a Turing model to be ready to pass to muse.

The model should be conditioned on the variables which comprise the "data", and all other variables should be unconditioned. params should give the variable names of the parameters which MUSE will estimate. All other non-conditioned and non-params variables will be considered the latent space. E.g.,

@model function demo()
    σ ~ Normal(0, 3)
    z ~ MvNormal(zeros(512), exp(σ/2))
    w ~ MvNormal(z, 1)
    x ~ MvNormal(w, 1)
    y ~ MvNormal(w, 2)
    (;σ,z,w,x,y)
end
truth = demo()()
model = demo() | (;truth.x, truth.y)
prob = TuringMuseProblem(model, params=(:σ,))

The autodiff parameter should be either MuseInference.ForwardDiffBackend() or MuseInference.ZygoteBackend(), specifying which library to use for automatic differenation.

Note

You can also call muse, etc... directly on the model, e.g. muse(model, (σ=1,), ...), in which case the parameter names params will be read from the keys of the starting point.